snap,systemd: refactor fstype for squashfs detection
The fstype for the squashfs can vary depending on the system in
use. On a regular system the kernel module "squashfs" is used,
however in lxd it can be "fuse.squashfuse" or "fuse.snapfuse".

Do the detection what to use in one central place.
