+++ overlord/snapstate/backend/mountunit_test.go
@@ -31,6 +31,7 @@ import (
 	"github.com/snapcore/snapd/overlord/snapstate/backend"
 	"github.com/snapcore/snapd/progress"
 	"github.com/snapcore/snapd/snap"
+	"github.com/snapcore/snapd/snap/squashfs"
 	"github.com/snapcore/snapd/systemd"
 	"github.com/snapcore/snapd/testutil"
 )
@@ -62,7 +63,7 @@ func (s *mountunitSuite) TearDownTest(c *C) {
 }
 
 func (s *mountunitSuite) TestAddMountUnit(c *C) {
-	restore := systemd.MockUseFuse(false)
+	restore := squashfs.MockUseFuse(false)
 	defer restore()
 
 	info := &snap.Info{
